+++ python/cuml/dask/common/__init__.py
@@ -13,5 +13,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-from cuml.dask.core import *
