+++ python/cuml/dask/linear_model/linear_regression.py
@@ -14,13 +14,16 @@
 # limitations under the License.
 #
 
+from cuml.dask.common.spmg import new_ipc_thread
+from cuml.dask.common.utils import parse_host_port
+from cuml.dask.common.utils import device_of_devicendarray
+from cuml.dask.common.utils import build_host_dict
+
 import cudf
 import dask_cudf
 import numpy as np
 
 from dask import delayed
-from cuml.dask.core import new_ipc_thread, parse_host_port
-from cuml.dask.core import device_of_devicendarray, build_host_dict
 from dask.distributed import wait, default_client
 from math import ceil
 from numba import cuda
