+++ python/cuml/dask/neighbors/nearest_neighbors.py
@@ -14,8 +14,9 @@
 #
 # from dask.core import *
 
-from cuml.dask.core import new_ipc_thread, parse_host_port
-from cuml.dask.core import device_of_devicendarray, build_host_dict
+from cuml.dask.common.spmg import new_ipc_thread
+from cuml.dask.common.utils import parse_host_port
+from cuml.dask.common.utils import device_of_devicendarray, build_host_dict
 
 from cuml.neighbors import NearestNeighbors as cumlKNN
